Python - Src packaged renamed to "pybookeeping" setup.py updated
The src packaged was renamed because this is an installable module, and
with the src package, the name of the package inside site-packages after
installation was called "src" which is misleading.

Giving a name to this project makes it pip-8 compliant and also gives a
name under site-packages

setup.py was updated and installed on the developer machine.